requirements.txt
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.29.0
